docs: update PR plan after modularization
